improve: move lyrics alignment logic
