+++ src/modules/lyricsAligner.js
